old/old1/stt/cpu_ws/requirements.txt
fastapi
uvicorn
openai-whisper
websockets
